fix: singleton mongo connection for Agenda
