+++ src/app/lib/job-handler/index.ts
@@ -1,27 +1,9 @@
-import { MongoBackend } from '@agendajs/mongo-backend';
 import { Agenda, Job } from 'agenda';
 import { GraphQLError } from 'graphql';
-import {
-  communityImportTask,
-  type CommunityImportJobArg,
-} from '~/graphql/schema/community/import';
-import {
-  batchPropertyModifyTask,
-  type BatchPropertyModifyJobArg,
-} from '~/graphql/schema/property/batch-modify';
-import { Logger } from '~/lib/logger';
-import mongoClient from '~/lib/mongo-client';
+import { getAgenda, type AgendaJobMap } from './agenda';
 import { JobEntry } from './job-entry';
-import { publishJobStatus } from './publish-job-status';
 
 export { JobEntry } from './job-entry';
-
-const logger = Logger('lib/job-handler');
-
-interface AgendaJobMap {
-  communityImport: CommunityImportJobArg;
-  batchPropertyModify: BatchPropertyModifyJobArg;
-}
 
 /**
  * Helper class for queuing long running job
@@ -32,42 +14,12 @@ export class JobHandler {
   constructor(private agenda: Agenda) {}
 
   /**
-   * Initialize Agenda task scheduler using the underlying database
+   * Initialize Agenda task scheduler
    *
    * @returns A JobHandler instance
    */
   static async init() {
-    const agenda = new Agenda({
-      backend: new MongoBackend({
-        // Reuse the single global instance of the mongo database
-        mongo: mongoClient.db(),
-      }),
-      /**
-       * It would be nice to removeOnComplete by default, but we need to handle
-       * the case when the subscription for job progress happens after the job
-       * has already been completed. So we leave this false, so we can manage
-       * the job document manually
-       */
-      removeOnComplete: false,
-    });
-
-    agenda.on('error', (err) => {
-      logger.error('Agenda error:', err);
-    });
-
-    /** There isn't much value to let UI know that a job has started */
-    // this._agenda.on('start', (job) => {
-    //   publishJobStatus(job);
-    // });
-    agenda.on('complete', async (job) => {
-      // This will be called whether the job succeed or fail
-      publishJobStatus(job);
-    });
-
-    agenda.define('communityImport', communityImportTask);
-    agenda.define('batchPropertyModify', batchPropertyModifyTask);
-
-    await agenda.start();
+    const agenda = await getAgenda();
     return new JobHandler(agenda);
   }
